requirements.txt
streamlit
pymongo
requests
python-dotenv
gemini
databricks-cli
